# src/shopping-admin/shopping-admin-consumer/docs/swagger.yaml
basePath: /api/shopping/admin
definitions:
...
  Request-any:
    properties:
      pageIndex:
        description: 分页索引
        example: 1
        minimum: 1
        type: integer
      pageSize:
        description: 分页长度
        example: 10
        maximum: 100
        minimum: 10
        type: integer
      params:
        description: 请求查询参数
    required:
    - pageIndex
    - pageSize
    type: object
...
  Request-shopping_admin_BusinessParam:
    properties:
      pageIndex:
        description: 分页索引
        example: 1
        minimum: 1
        type: integer
      pageSize:
        description: 分页长度
        example: 10
        maximum: 100
        minimum: 10
        type: integer
      params:
        allOf:
        - $ref: '#/definitions/shopping_admin.BusinessParam'
        description: 请求查询参数
    required:
    - pageIndex
    - pageSize
    type: object
...
  Request-shopping_admin_ProductSpecParam:
    properties:
      pageIndex:
        description: 分页索引
        example: 1
        minimum: 1
        type: integer
      pageSize:
        description: 分页长度
        example: 10
        maximum: 100
        minimum: 10
        type: integer
      params:
        allOf:
        - $ref: '#/definitions/shopping_admin.ProductSpecParam'
        description: 请求查询参数
    required:
    - pageIndex
    - pageSize
    type: object
  Response:
    properties:
      code:
        default: 0
        description: 结果状态码
        type: integer
      data:
        description: 结果
      msg:
        default: 请求失败
        description: 状态信息
        type: string
    type: object
...
  shopping_admin.Business:
    properties:
      BackgroundImg:
        description: 背景图地址
        type: string
      ChargeOpen:
        default: 2
        description: 扣款（1：开启，2：关闭[默认]）
        maximum: 2
        minimum: 1
        type: integer
      Name:
        description: 商家名称
        type: string
      SettlementWithdrawalDay:
        description: 订单入账后X天可提现（0表示马上可提现）
        type: integer
      Views:
        description: 浏览次数
        type: integer
      address:
        description: 联系人地址
        type: string
      agentId:
        description: 所属代理商id
        type: string
      aloneExamine:
        default: 2
        description: 提现至余额单独设置（1：开启，2：关闭[默认]）
        maximum: 2
        minimum: 1
        type: integer
      arrearsMoney:
        description: 欠费金额
        type: number
      balanceCashOut:
        default: 1
        description: 余额提现是否审核（1：需要审核[默认]，2：无需审核）
        maximum: 2
        minimum: 1
        type: integer
      bankCardHolder:
        description: 银行卡持卡人姓名
        type: string
      bankCardNo:
        description: 银行卡号
        type: string
      bankId:
        description: 银行ID
        type: string
      businessId:
        description: 商家id
        type: string
      businessSettlement:
        default: 1
        description: 商家结算费率（1按平台[默认]、2自定义）
        maximum: 2
        minimum: 1
        type: integer
      cashOutExamine:
        default: 1
        description: 可提现金额至余额提现审核（1：需要[默认]，2：不需要）
        maximum: 2
        minimum: 1
        type: integer
      categoryIds:
        description: 上传可选分类（JSON）
        type: string
      chargeData:
        description: m_1：1月，m_2：2月......（title：扣款项，money：扣款金额）
        type: string
      chargeStartTime:
        description: 扣款起始日期
        type: string
      cid:
        description: 合伙人id
        type: string
      cityDeliveryRule:
        description: 同城配送单独设置计费规则
        type: string
      contacts:
        description: 联系人
        type: string
      description:
        description: 描述
        type: string
      expressType:
        description: 数据展示类型 2:是自提，3:是核销json格式
        type: string
      franchiseeId:
        description: 入驻ID（非商家入驻为0，是商家入驻为主键id）
        type: string
      goodsExamine:
        default: 1
        description: 商品审核（1：需要[默认]，2：不需要）
        maximum: 2
        minimum: 1
        type: integer
      isCityDeliveryAlone:
        default: 2
        description: 同城配送配送费，商家单独设置（1：开启，2：关闭[默认]）
        maximum: 2
        minimum: 1
        type: integer
      isOpenData:
        default: 0
        description: 是否开始单独数据展示(0:否[默认]，1：是)
        maximum: 1
        minimum: 0
        type: integer
      isShowV1:
        description: 是否展示到V1样式（1：是，2：否）
        type: integer
      logo:
        description: logo地址
        type: string
      merchantId:
        description: 头条商户id
        type: string
      money:
        description: 代理后启用，商家余额
        type: number
      platformAppId:
        description: 平台应用
        type: string
      remark:
        description: 备注
        type: string
      salesData:
        description: 销售数据展示权限
        type: string
      salesVolume:
        description: 销售额
        type: string
      settlement:
        description: 默认结算率
        type: number
      settlementMoney:
        description: 商家累积结算金额
        type: number
      showBusinessQualification:
        default: "1"
        description: 商家资质;1：展示[默认]，2隐藏
        type: string
      sortBy:
        description: 商家排序
        type: integer
      telephone:
        description: 联系人电话
        type: string
      wechat:
        description: 联系人微信
        type: string
    required:
    - Name
    - businessId
    - description
    - logo
    type: object
...
  shopping_admin.BusinessParam:
    type: object
...
  shopping_admin.ProductCategory:
    properties:
      name:
        description: 名称
        type: string
      parentId:
        description: 上级分类
        type: string
      productCategoryId:
        description: 商品分类id
        type: string
      profitSharing:
        description: 普通店铺分佣比例
        type: number
      selfRate:
        description: 自营店铺分佣比例
        type: number
      sortBy:
        description: 排序
        type: integer
    required:
    - name
    - parentId
    - productCategoryId
    type: object
...
  shopping_admin.ProductSpec:
    properties:
      name:
        description: 规格名称
        type: string
      options:
        description: 可选项
        type: string
      productCategoryId:
        description: 绑定分类
        type: string
      sortBy:
        description: 排序
        type: integer
      specId:
        description: 商品规格id
        type: string
    required:
    - name
    - options
    - productCategoryId
    type: object
  shopping_admin.ProductSpecParam:
    type: object
...
info:
  contact:
    email: 826422592@qq.com
    name: 杨大琼
    url: http://localhost
  description: swagger go API文档
  license:
    name: Apache 2.0
  title: 商城后台管理服务提供者接口文档
paths:
...
  /business:
    delete:
      description: 主要用于商家删除
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 删除参数
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 删除商家
      tags:
      - 商家管理
    post:
      description: 主要用于商家新增
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 新增商家参数
        in: body
        name: business
        required: true
        schema:
          $ref: '#/definitions/shopping_admin.Business'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 创建商家
      tags:
      - 商家管理
    put:
      description: 主要用于商家修改
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 修改商家参数
        in: body
        name: business
        required: true
        schema:
          $ref: '#/definitions/shopping_admin.Business'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 修改商家
      tags:
      - 商家管理
  /business/findById/{id}:
    get:
      description: 根据id查询商家内容详情
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 商家id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 根据id查询商家内容详情
      tags:
      - 商家管理
  /business/findPageList:
    post:
      description: 主要用于查询商家分页列表
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 查询参数
        in: body
        name: pageInfo
        required: true
        schema:
          $ref: '#/definitions/Request-shopping_admin_BusinessParam'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 商家分页列表
      tags:
      - 商家管理
  /businessCircle:
    delete:
      description: 主要用于商圈删除
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 商圈是删除参数
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 删除商圈
      tags:
      - 商圈管理
  /businessCircle/findById/{id}:
    get:
      description: 根据id查询商圈内容详情
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 商圈id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 根据id查询商圈内容详情
      tags:
      - 商圈管理
  /businessCircle/findPageList:
    post:
      description: 主要用于查询商圈分页列表
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 查询参数
        in: body
        name: pageInfo
        required: true
        schema:
          $ref: '#/definitions/Request-any'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 商圈分页列表
      tags:
      - 商圈管理
...
  /category:
    delete:
      description: 主要用于商品分类删除
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 删除参数
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 删除商品分类
      tags:
      - 商品分类
    post:
      description: 主要用于商品分类新增
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 新增商品分类参数
        in: body
        name: productCategory
        required: true
        schema:
          $ref: '#/definitions/shopping_admin.ProductCategory'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 创建商品分类
      tags:
      - 商品分类
    put:
      description: 主要用于商品分类修改
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 修改商品分类参数
        in: body
        name: productCategory
        required: true
        schema:
          $ref: '#/definitions/shopping_admin.ProductCategory'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 修改商品分类
      tags:
      - 商品分类
  /category/findListByParentId/{parentId}:
    get:
      description: 主要用于根据商品父id查询下级分类
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 分类父id
        in: path
        name: parentId
        required: true
        type: string
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 根据商品父id查询下级分类
      tags:
      - 商品分类
  /category/findTreeByParentId/{parentId}:
    get:
      description: 主要用于根据商品父id查询所有子分类
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 分类父id
        in: path
        name: parentId
        required: true
        type: string
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 根据商品父id查询所有子分类【树形结构】
      tags:
      - 商品分类
...
  /spec:
    delete:
      description: 主要用于商品规格删除
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 删除参数
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 删除商品规格
      tags:
      - 商品规格管理
    post:
      description: 主要用于商品规格新增
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 新增商品规格参数
        in: body
        name: spec
        required: true
        schema:
          $ref: '#/definitions/shopping_admin.ProductSpec'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 创建商品规格
      tags:
      - 商品规格管理
    put:
      description: 主要用于商品规格修改
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 修改商品规格参数
        in: body
        name: spec
        required: true
        schema:
          $ref: '#/definitions/shopping_admin.ProductSpec'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 修改商品规格
      tags:
      - 商品规格管理
  /spec/findById/{id}:
    get:
      description: 根据id查询商品规格内容详情
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 商品规格id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 根据id查询商品规格内容详情
      tags:
      - 商品规格管理
  /spec/findList/{id}:
    get:
      description: 根据productCategoryId查询规格列表
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 商品分类id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 根据productCategoryId查询规格列表
      tags:
      - 商品规格管理
  /spec/findPageList:
    post:
      description: 主要用于查询商品规格分页列表
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 查询参数
        in: body
        name: pageInfo
        required: true
        schema:
          $ref: '#/definitions/Request-shopping_admin_ProductSpecParam'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 商品规格分页列表
      tags:
      - 商品规格管理
...
swagger: "2.0"